Refactor team mixins
