--- gratipay/models/team/mixins/__init__.py
@@ -1,7 +1,0 @@
-from .available import AvailableMixin as Available
-from .closing import ClosingMixin as Closing
-from .membership import MembershipMixin as Membership
-from .takes import TakesMixin as Takes
-from .tip_migration import TipMigrationMixin as TipMigration
-
-__all__ = ['Available', 'Closing', 'Membership', 'Takes', 'TipMigration']
